Move accesspoint creation back
